requirements.txt
streamlit>=1.54.0
requests>=2.32.5
python-dotenv>=1.2.1
huggingface_hub>=1.4.1
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.7.0
reportlab>=4.0.0
pillow>=10.0.0